boardView repository 추가
